Use popen3 instead of childprocess to run commands
childprocess doesn't allow us to get access to stdout and stderr
separately and also as combined output, but we can accomplish this with
popen3. This requires using popen3 in an advanced way. I admit I'm not
sure I understand everything that is going on here but it seems to work
well enough for my needs.